% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchSenadoProposal}
\alias{fetchSenadoProposal}
\title{Fetch datailed information of a proposal from the Senate}
\usage{
fetchSenadoProposal(id)
}
\arguments{
\item{id}{The id of the proposition as recorded in Senado}
}
\value{
Dataframe with detailed information from a legislation proposal
}
\description{
Retorna dataframe com os dados detalhados da proposição, incluindo número, ementa, tipo e data de apresentação.
Ao fim, a função retira todos as colunas que tenham tipo lista para uniformizar o dataframe.
}
\examples{
\dontrun{
df = fetchSenadoProposal(91341)
}
}
