% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codigos.R
\name{fetchSenadoProposalDetails}
\alias{fetchSenadoProposalDetails}
\title{Recupera e processa dados de uma proposição}
\usage{
fetchSenadoProposalDetails(id)
}
\arguments{
\item{id}{ID da proposição}
}
\value{
Dataframe com os dados de proposições
}
\description{
A partir de um id, retorna os dados de uma proposição
}
\examples{
\dontrun{
df = fetchSenadoProposalDetails(91341)
}
}
